% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_functions.R
\name{prelec}
\alias{prelec}
\title{Prelec Two Parameter Recalibration Function}
\usage{
prelec(p, alpha, beta)
}
\arguments{
\item{p}{a numeric vector of probabilities to be Prelec-adjusted. Must only
contain values in [0,1].}

\item{alpha}{numeric, must be > 0, \eqn{\alpha} in Prelec two parameter function.}

\item{beta}{numeric, must be > 0, \eqn{\beta} in Prelec two parameter function.}
}
\description{
Prelec Two Parameter Recalibration Function
}
