% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boldness_recalib.R
\name{brcal}
\alias{brcal}
\title{Boldness Recalibration for Binary Events}
\usage{
brcal(
  x,
  y,
  t = 0.95,
  Pmc = 0.5,
  tau = FALSE,
  event = 1,
  start_at_MLEs = TRUE,
  x0 = NULL,
  lb = NULL,
  ub = NULL,
  maxeval = 500,
  maxtime = NULL,
  xtol_rel = 1e-06,
  print_level = 3,
  opts = NULL
)
}
\arguments{
\item{x}{a numeric vector of probabilities to be LLO-adjusted. Must only
contain values in [0,1].}

\item{y}{a numeric vector of outcomes corresponding to probabilities in \code{x}.
Must only contain 0 or 1.}

\item{Pmc}{The prior model probability for the calibrated model \eqn{M_c}.}

\item{event}{Value in \code{y} that represents an "event".  Default value is 1.}

\item{start_at_MLEs}{Logical. If TRUE, the optimizer will start at
\eqn{\delta} and \eqn{x_0} = the maximum likelihood estimates for
\eqn{\delta} and \eqn{\gamma}. Otherwise, the user must specify
\eqn{\delta} and \eqn{x_0}.}

\item{x0}{Vector with starting locations for \eqn{\delta} and \eqn{\gamma}.  This argument is ignored when start_at_MLEs = TRUE.}

\item{lb}{Vector with lower bounds for \eqn{\delta} and \eqn{\gamma}.}

\item{ub}{Vector with upper bounds for \eqn{\delta} and \eqn{\gamma}.}

\item{opts}{list with options to be passed to \code{nloptr}.  See the documentation for \code{nloptr()} for details.}
}
\description{
Boldness Recalibration for Binary Events
}
