% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_functions.R
\name{LLO}
\alias{LLO}
\title{Linear Log Odds (LLO) Recalibration Function}
\usage{
LLO(x, delta, gamma)
}
\arguments{
\item{x}{a numeric vector of probabilities to be LLO-adjusted. Must only
contain values in [0,1].}

\item{delta}{numeric, must be > 0, parameter \eqn{\delta} in LLO
recalibration function.}

\item{gamma}{numeric, parameter \eqn{\gamma} in LLO recalibration function.}
}
\value{
Vector of LLO-adjusted probabilities via specified \eqn{\delta} and
\eqn{\gamma}.
}
\description{
LLO-adjust predicted probabilities based on specified \eqn{\delta} and
\eqn{\gamma}.
}
\references{
Turner, B., Steyvers, M., Merkle, E., Budescu, D., and Wallsten,
T. (2014) Forecast aggregation via recalibration, \emph{Machine Learning}
95, 261–289.

Gonzalez, R., and Wu, G. (1999), On the shape of probability
weighting function, \emph{Cognitive Psychology} 38, 129–66.
}
