% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_functions.R
\name{LLO}
\alias{LLO}
\title{Linear Log Odds (LLO) Recalibration Function}
\usage{
LLO(p, delta, gamma)
}
\arguments{
\item{p}{a numeric vector of probabilities to be LLO-adjusted}

\item{delta}{numeric, shift parameter (on log odds scale), must be > 0.}

\item{gamma}{numeric, scale parameter (on log odds scale).}
}
\value{
The LLO-adjusted vector of probabilities (ADD NOTATION FROM PAPER? OR KEEP IN DETAILS?)
}
\description{
Description goes here.
}
\details{
Details go here. NEED TO CITE TURNER?
}
