% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_functions.R
\name{LLO}
\alias{LLO}
\title{Linear Log Odds (LLO) Recalibration Function}
\usage{
LLO(x, delta, gamma, ...)
}
\arguments{
\item{x}{a numeric vector of probabilities to be LLO-adjusted. Must only
contain values in [0,1].}

\item{delta}{numeric, must be > 0, parameter \eqn{\delta} in LLO
recalibration function.}

\item{gamma}{numeric, parameter \eqn{\gamma} in LLO recalibration function.}

\item{...}{Additional arguments for internal use only.}
}
\value{
The LLO-adjusted vector of probabilities (ADD NOTATION FROM PAPER? OR
KEEP IN DETAILS?)
}
\description{
Description goes here.
}
\details{
Details go here. NEED TO CITE TURNER/Gonzalez & Wu? Consider removing
functionality for lists/matrices (any other types?).
}
