% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_functions.R
\name{llo_lrt}
\alias{llo_lrt}
\title{Likelihood Ratio Test for Calibration [NEED MORE]}
\usage{
llo_lrt(x, y, optim_details = TRUE, ...)
}
\arguments{
\item{x}{a numeric vector of probabilities to be LLO-adjusted. Must only
contain values in [0,1].}

\item{y}{a numeric vector of outcomes corresponding to probabilities in \code{x}.
Must only contain 0 or 1.}

\item{optim_details}{Logical.  If \code{TRUE}, the list returned by \code{optim()} when
minimizing the negative log likelihood is also returned by \code{llo_lrt()}.}

\item{...}{Additional arguments to be passed to \code{optim()}.}
}
\value{
A list with the following attributes: \item{\code{test_stat}}{The
test statistic from the likelihood ratio test formed as FILL IN}
\item{\code{pval}}{The p-value from the likelihood ratio test.}
\item{\code{mles}}{Maximum likelihood estimates for \eqn{\delta} and \eqn{\gamma}.}
\item{\code{optim_details}}{If \code{optim_details = TRUE}, the list returned by
\code{optim()} when minimizing the negative log likelihood, includes convergence
information, number of iterations, and achieved negative log likelihood
value and MLEs.}
}
\description{
Likelihood Ratio Test for Calibration [NEED MORE]
}
