% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_functions.R
\name{mle_recal}
\alias{mle_recal}
\title{Title}
\usage{
mle_recal(x, y, probs_only = TRUE, optim_details = TRUE, ...)
}
\arguments{
\item{x}{a numeric vector of probabilities to be LLO-adjusted. Must only
contain values in [0,1].}

\item{y}{a numeric vector of outcomes corresponding to probabilities in \code{x}.
Must only contain 0 or 1.}

\item{probs_only}{blah}

\item{optim_details}{Logical.  If \code{TRUE}, the list returned by \code{optim()} when
minimizing the negative log likelihood is also returned by \code{llo_lrt()}.}

\item{...}{Additional arguments to be passed to \code{optim()}.}
}
\description{
Title
}
