% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_params}
\alias{plot_params}
\title{Draw image plot of posterior model probability surface.}
\usage{
plot_params(
  x,
  y,
  z = NULL,
  t_levels = c(0.8, 0.9),
  Pmc = 0.5,
  event = 1,
  k = 100,
  dlim = c(1e-04, 5),
  glim = c(1e-04, 5),
  zlim = c(0, 1),
  return_z = TRUE,
  thin_to = NULL,
  thin_percent = NULL,
  thin_by = NULL,
  contours_only = FALSE,
  main = "Posterior Model Probability of Calibration",
  xlab = "delta",
  ylab = "gamma",
  legend.lab = "",
  drawlabels = TRUE,
  contour_color = "white",
  labcex = 0.6,
  lwd = 1,
  ...
)
}
\arguments{
\item{x}{a numeric vector of probabilities to be LLO-adjusted. Must only
contain values in [0,1].}

\item{y}{a numeric vector of outcomes corresponding to probabilities in \code{x}.
Must only contain 0 or 1.}

\item{z}{Matrix returned by previous call to \code{plot_params()} containing
posterior model probabilities across k\eqn{\times}k grid of \eqn{\delta}
and \eqn{\gamma}. Assumes \code{z} was constructed using the same \code{k}, \code{dlim},
and \code{glim} as the current function call.}

\item{t_levels}{Vector of desired level(s) of calibration at which to plot
contours.}

\item{Pmc}{The prior model probability for the calibrated model \eqn{M_c}.}

\item{event}{Value in \code{y} that represents an "event".  Default value is 1.}

\item{k}{The number of uniformly spaced \eqn{\delta} and \eqn{\gamma} values
used to construct the k\eqn{\times}k grid.}

\item{dlim}{Vector with bounds for \eqn{\delta}, must be finite.}

\item{glim}{Vector with bounds for \eqn{\gamma}, must be finite.}

\item{zlim}{Vector with bounds for posterior probability of calibration, must
be in [0,1].}

\item{return_z}{Logical.  If \code{TRUE}, the matrix of posterior model
probabilities across the specified k\eqn{\times}k grid of \eqn{\delta} and
\eqn{\gamma} will be returned.}

\item{thin_to}{When non-null, the observations in (x,y) are randomly sampled
without replacement to form a set of size \code{thin_to}.}

\item{thin_percent}{When non-null, the observations in (x,y) are randomly
sampled without replacement to form a set that is \code{thin_percent}\\% of the
original size of (x,y).}

\item{thin_by}{When non-null, the observations in (x,y) are thinned by
selecting every \code{thin_by} observation.}

\item{contours_only}{Logical.  If \code{TRUE}, only the contours at the specified
\code{t_levels} will be plotted with no color for the posterior model
probability across the k\eqn{\times}k grid of \eqn{\delta} and
\eqn{\gamma}.}

\item{main}{Plot title.}

\item{xlab}{Label for x-axis.}

\item{ylab}{Label for x-axis.}

\item{drawlabels}{Logical.  If \code{TRUE}, the contour labels will be plotted
with the corresponding contours passed to \code{contour()}.}

\item{contour_color}{Color of the contours and their labels passed to
\code{contour()}.}

\item{labcex}{Size of contour labels passed to \code{contour()}.}

\item{lwd}{Linewidth of contours passed to \code{contour()}.}

\item{...}{Additional arguments to be passed to \code{image}, \code{image.plot}, and
\code{contour}.}
}
\value{
If \code{return_z = TRUE}, a list with the following attributes is
returned: \item{\code{z}}{Matrix containing posterior model probabilities
across k\eqn{\times}k grid of uniformly spaced values of \eqn{\delta} and \eqn{\gamma}
in the specified ranges \code{dlim} and \code{glim}, respectively.}
\item{\code{dlim}}{Vector with bounds for
\eqn{\delta} used to construct z.}
\item{\code{glim}}{Vector with bounds for \eqn{\gamma} used to construct
z.}
\item{\code{k}}{The number of uniformly spaced \eqn{\delta} and \eqn{\gamma} values used to construct
z}
}
\description{
Function to plot the posterior model probability of the given set of
probabilities, \code{x}, after LLO-adjustment via a grid of uniformly spaced set
of \eqn{\delta} and \eqn{\gamma} values with optional contours.
}
