% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_params}
\alias{plot_params}
\title{Title}
\usage{
plot_params(
  z,
  len.out = 100,
  lower = c(1e-04, -2),
  upper = c(5, 2),
  cont_levels = c(0.8, 0.9),
  sub = "",
  zlim = c(0, 1),
  ttle_extra = "",
  ttle = "Posterior Model Probability of Calibration",
  contours_only = FALSE,
  add = FALSE,
  contour_color = "white",
  leg_lab = "",
  drawlabels = TRUE,
  xlab = "delta",
  ylab = "gamma",
  lwd = 1,
  labcex = 0.6,
  legend.args = list(las = 180),
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Title
}
